AuthController update

+++ src/main/resources/db/migration/V1__Init_DB_tables.sql
@@ -11,6 +11,6 @@ CREATE TABLE short_urls (
     date_of_creating DATE,
     date_of_expiring DATE,
     count_of_transition BIGINT,
-    user_id INT NOT NULL,
+    user_id BIGSERIAL NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users(id)
 );
